chore: remove survey package
