+++ cmd/project/create_samples.go
@@ -21,7 +21,6 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/slackapi/slack-cli/internal/experiment"
 	"github.com/slackapi/slack-cli/internal/iostreams"
 	"github.com/slackapi/slack-cli/internal/pkg/create"
 	"github.com/slackapi/slack-cli/internal/shared"
@@ -67,21 +66,13 @@ func promptSampleSelection(ctx context.Context, clients *shared.ClientFactory, s
 	sortedRepos := sortRepos(filteredRepos)
 	selectOptions := make([]string, len(sortedRepos))
 	for i, r := range sortedRepos {
-		if !clients.Config.WithExperimentOn(experiment.Huh) {
-			selectOptions[i] = fmt.Sprint(i+1, ". ", r.Name)
-		} else {
-			selectOptions[i] = r.Name
-		}
+		selectOptions[i] = r.Name
 	}
 
 	var selectedTemplate string
 	selection, err = clients.IO.SelectPrompt(ctx, "Select a sample to build upon:", selectOptions, iostreams.SelectPromptConfig{
 		Description: func(value string, index int) string {
-			desc := sortedRepos[index].Description
-			if !clients.Config.WithExperimentOn(experiment.Huh) {
-				desc += "\n  https://github.com/" + sortedRepos[index].FullName
-			}
-			return desc
+			return sortedRepos[index].Description
 		},
 		Flag:     clients.Config.Flags.Lookup("template"),
 		Help:     fmt.Sprintf("Guided tutorials can be found at %s", style.LinkText("https://docs.slack.dev/samples")),
