+++ internal/experiment/experiment_test.go
@@ -25,7 +25,6 @@ func Test_Includes(t *testing.T) {
 	require.Equal(t, true, Includes(Experiment(Placeholder)))
 
 	// Test expected experiments
-	require.Equal(t, true, Includes(Experiment("huh")))
 	require.Equal(t, true, Includes(Experiment("lipgloss")))
 
 	// Test invalid experiment
