+++ internal/style/theme_test.go
@@ -18,7 +18,6 @@ import (
 	"runtime"
 	"testing"
 
-	"github.com/AlecAivazis/survey/v2/core"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -143,29 +142,6 @@ func TestChevron(t *testing.T) {
 				isStyleEnabled = prev
 			}()
 			assert.Equal(t, tc.expected, Chevron())
-		})
-	}
-}
-
-func TestSurveyIcons(t *testing.T) {
-	tests := map[string]struct {
-		styleEnabled bool
-	}{
-		"styles are not enabled": {
-			styleEnabled: false,
-		},
-		"styles are enabled": {
-			styleEnabled: true,
-		},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			core.DisableColor = false
-			isStyleEnabled = tc.styleEnabled
-
-			_ = SurveyIcons()
-			assert.NotEqual(t, tc.styleEnabled, core.DisableColor)
 		})
 	}
 }
